# Avoid out of range array access in JBIG2 decoder