Fixes #5663. Chrome rendering times are down from 4 seconds to <1.5s.

Benchmark with document from #2909:

browser        | Count | Baseline(ms) | Current(ms) |   +/- |     %  | Result(P<.05)
-------------- | ----- | ------------ | ----------- | ----- | ------ | -------------
chrome         |   180 |         3215 |         997 | -2218 | -69.00 |        faster
firefoxNightly |   180 |         1156 |        1139 |   -18 |  -1.53 |