@yurydelendik 2015-01-26
/botio-windows test