@CodingFabian 2015-01-26
I was just looking at the code, and I wondered if you have tested unrolling that magic or yourself

```
contextLabel = ((contextLabel & OLD_PIXEL_MASK) << 1) |
                         (j + 3 < width ? row2[j + 3] << 11 : 0) |
                       (j + 4 < width ? row1[j + 4] << 4 : 0) | pixel;
```

could be
```
if (contextLabel) {
  // the only way the & shift will become 0 is if contextLabel is 0
  contextLabel = (contextLabel & OLD_PIXEL_MASK) << 1);
  continue;
} 
if (j + 3 < width) {
 contextLabel = row2[j + 3] << 11;
 if (contextLabel) continue;
}
if (j + 4 < width) {
 contextLabel = row1[j + 4] << 4;
 if (contextLabel) continue;
}
contextLabel = pixel;
```