@fkaelberer 2015-01-27
@CodingFabian Before I committed this fix, I have tested similar code to avoid the branching in the `j + x < width` branching in the loop, but the performance gain was below 1 or 2% for ~15 additional lines of code. But I preferred simplicity/readability over the marginal speed gain. And don't confuse `|` with `||`: you can't continue early if contextLabel is != 0 :-]